Unify webpack config
